refactor: splitting github utils into smaller files
