+++ internal/pkg/util/github/repo.go
@@ -1,0 +1,19 @@
+package github
+
+import (
+	"context"
+
+	"github.com/google/go-github/v42/github"
+)
+
+func (c *Client) CreateRepo() error {
+	repo := &github.Repository{
+		Name: &c.Repo,
+	}
+
+	_, _, err := c.Repositories.Create(context.TODO(), "", repo)
+	if err != nil {
+		return err
+	}
+	return nil
+}
